test(datavisualisations): move chart tests into nested directories
